usid: move _sid_req to iface/usid.c
In preparation for other users of the sid ubridge socket, rename
_sid_req to usid_req and move it to iface/usid.c. usid_req() needs an
additional parameter, prefix, to use for logging. There are no other
functional changes.
