+++ src/tools/usid/usid.c
@@ -20,7 +20,6 @@
 #include "base/common.h"
 
 #include "base/buffer.h"
-#include "base/comms.h"
 #include "base/util.h"
 #include "log/log.h"
 #include "iface/usid.h"
@@ -30,7 +29,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <sys/sysmacros.h>
-#include <unistd.h>
 
 #define LOG_PREFIX                      "usid"
 
@@ -54,83 +52,10 @@
 #define KEY_SID_MINOR                   "SID_MINOR"
 #define KEY_SID_RELEASE                 "SID_RELEASE"
 
-typedef int (*sid_req_data_fn_t) (struct buffer *buf, void *data);
-
 struct args {
 	int argc;
 	char **argv;
 };
-
-static int _sid_req(usid_cmd_t cmd, uint64_t status, sid_req_data_fn_t data_fn, void *data_fn_arg, struct buffer **resp_buf)
-{
-	int socket_fd = -1;
-	struct buffer *buf = NULL;
-	ssize_t n;
-	int r = -1;
-
-	if (!(buf = buffer_create(BUFFER_TYPE_LINEAR, BUFFER_MODE_SIZE_PREFIX, 0, 1, 0, &r))) {
-		log_error_errno(LOG_PREFIX, r, "Failed to create request buffer");
-		goto out;
-	}
-
-	if (!buffer_add(buf,
-	&((struct usid_msg_header) {
-	.status = status,
-	.prot = USID_PROTOCOL,
-	.cmd = cmd
-}), USID_MSG_HEADER_SIZE, &r))
-	goto out;
-
-	if (data_fn && (data_fn(buf, data_fn_arg) < 0)) {
-		log_error(LOG_PREFIX, "Failed to add data to request.");
-		goto out;
-	}
-
-	if ((socket_fd = comms_unix_init(USID_SOCKET_PATH, USID_SOCKET_PATH_LEN, SOCK_STREAM | SOCK_CLOEXEC)) < 0) {
-		r = socket_fd;
-		if (r != -ECONNREFUSED)
-			log_error_errno(LOG_PREFIX, r, "Failed to initialize connection");
-		goto out;
-	}
-
-	if (buffer_write(buf, socket_fd, 0) < 0) {
-		log_error(LOG_PREFIX, "Failed to send request.");
-		goto out;
-	}
-
-	buffer_reset(buf);
-
-	for (;;) {
-		n = buffer_read(buf, socket_fd);
-		if (n > 0) {
-			if (buffer_is_complete(buf, NULL)) {
-				r = 0;
-				break;
-			}
-		} else if (n < 0) {
-			if (n == -EAGAIN || n == -EINTR)
-				continue;
-			log_error_errno(LOG_PREFIX, errno, "Failed to read response");
-			r = -EBADMSG;
-			break;
-		} else {
-			if (!buffer_is_complete(buf, NULL))
-				log_error(LOG_PREFIX, "Unexpected reponse end.");
-			break;
-		}
-	}
-out:
-	if (socket_fd >= 0)
-		close(socket_fd);
-
-	if (r < 0) {
-		if (buf)
-			buffer_destroy(buf);
-	} else
-		*resp_buf = buf;
-
-	return r;
-}
 
 static int _usid_cmd_active(struct args *args)
 {
@@ -144,7 +69,7 @@ static int _usid_cmd_active(struct args *args)
 
 	seqnum = util_env_get_ull(KEY_ENV_SEQNUM, 0, UINT64_MAX, &val) < 0 ? 0 : val;
 
-	if ((r = _sid_req(USID_CMD_VERSION, seqnum, NULL, NULL, &buf)) == 0) {
+	if ((r = usid_req(LOG_PREFIX, USID_CMD_VERSION, seqnum, NULL, NULL, &buf)) == 0) {
 		buffer_get_data(buf, (const void **) &hdr, &size);
 
 		if ((size >= (USID_MSG_HEADER_SIZE + USID_VERSION_SIZE)) &&
@@ -267,7 +192,7 @@ static int _usid_cmd_checkpoint(struct args *args)
 
 	seqnum = val;
 
-	if ((r = _sid_req(USID_CMD_CHECKPOINT, seqnum, _add_checkpoint_env_to_buf, args, &buf)) == 0) {
+	if ((r = usid_req(LOG_PREFIX, USID_CMD_CHECKPOINT, seqnum, _add_checkpoint_env_to_buf, args, &buf)) == 0) {
 		r = _print_env_from_buffer(buf);
 		buffer_destroy(buf);
 	}
@@ -305,7 +230,7 @@ static int _usid_cmd_scan(struct args *args)
 
 	seqnum = val;
 
-	if ((r = _sid_req(USID_CMD_SCAN, seqnum, _add_scan_env_to_buf, NULL, &buf)) == 0) {
+	if ((r = usid_req(LOG_PREFIX, USID_CMD_SCAN, seqnum, _add_scan_env_to_buf, NULL, &buf)) == 0) {
 		r = _print_env_from_buffer(buf);
 		buffer_destroy(buf);
 	}
@@ -334,7 +259,7 @@ static int _usid_cmd_version(struct args *args)
 	        SID_VERSION_MINOR,
 	        SID_VERSION_RELEASE);
 
-	if ((r = _sid_req(USID_CMD_VERSION, seqnum, NULL, NULL, &buf)) == 0) {
+	if ((r = usid_req(LOG_PREFIX, USID_CMD_VERSION, seqnum, NULL, NULL, &buf)) == 0) {
 		buffer_get_data(buf, (const void **) &hdr, &size);
 
 		if (size >= (USID_MSG_HEADER_SIZE + USID_VERSION_SIZE)) {
